frontend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
Pillow==10.3.0
supabase==2.4.1
python-dotenv==1.0.1
